requirements.txt
tensorflow>=2.10
numpy>=2.0
flask
pillow
scikit-learn
seaborn
